Speed up page loading by not waiting for visualization process
Rather than wait for the python process to start, load the
default view. Then used the promise returned by the vtkWeb.start(...)
call to ensure that the connection is complete before loading a
molecules 3D visualization. Thus the visualization session is
established while the user is looking a search results.